chore: rename CollentionDescription to CollectionDescription
